tools/dataset/requirements-train.txt
tensorflow-cpu>=2.16.1
numpy>=1.26.0
Pillow>=10.0.0
